Creating SRC and Results Folders
I created the SRC folder that holds the .py files and the Results folder that holds documentation of the results.
