+++ Results/pyscript.LGBM(classifier)-7-5-2021.txt
@@ -1,0 +1,60 @@
+=== Classifier model (full training set) ===
+
+LightGBM
+
+Time taken to build model: 6.82 seconds
+
+=== Stratified cross-validation ===
+=== Summary ===
+
+Correctly Classified Instances      125537               99.6539 %
+Incorrectly Classified Instances       436                0.3461 %
+Kappa statistic                          0.993 
+Mean absolute error                      0.1856
+Root mean squared error                  0.1892
+Relative absolute error                 37.3049 %
+Root relative squared error             37.9377 %
+Total Number of Instances           125973     
+
+=== Detailed Accuracy By Class ===
+
+                 TP Rate  FP Rate  Precision  Recall   F-Measure  MCC      ROC Area  PRC Area  Class
+                 0.998    0.005    0.996      0.998    0.997      0.993    0.999     0.999     normal
+                 0.995    0.002    0.997      0.995    0.996      0.993    0.999     0.999     anomaly
+Weighted Avg.    0.997    0.004    0.997      0.997    0.997      0.993    0.999     0.999     
+
+=== Confusion Matrix ===
+
+     a     b   <-- classified as
+ 67185   158 |     a = normal
+   278 58352 |     b = anomaly
+
+
+=== Re-evaluation on test set ===
+
+User supplied test set
+Relation:     KDDTest
+Instances:     unknown (yet). Reading incrementally
+Attributes:   42
+
+=== Summary ===
+
+Correctly Classified Instances       17867               79.2539 %
+Incorrectly Classified Instances      4677               20.7461 %
+Kappa statistic                          0.5978
+Mean absolute error                      0.3222
+Root mean squared error                  0.4067
+Total Number of Instances            22544     
+
+=== Detailed Accuracy By Class ===
+
+                 TP Rate  FP Rate  Precision  Recall   F-Measure  MCC      ROC Area  PRC Area  Class
+                 0.973    0.344    0.682      0.973    0.802      0.640    0.906     0.828     normal
+                 0.656    0.027    0.969      0.656    0.783      0.640    0.906     0.918     anomaly
+Weighted Avg.    0.793    0.164    0.845      0.793    0.791      0.640    0.906     0.879     
+
+=== Confusion Matrix ===
+
+    a    b   <-- classified as
+ 9445  266 |    a = normal
+ 4411 8422 |    b = anomaly
