+++ Results/pyscript.lightgbm(proto)-6-18-2021.txt
@@ -1,0 +1,77 @@
+=== Classifier model (full training set) ===
+
+lighgbm is running
+
+Time taken to build model: 13.75 seconds
+
+=== Stratified cross-validation ===
+=== Summary ===
+
+Correctly Classified Instances      125850               99.9024 %
+Incorrectly Classified Instances       123                0.0976 %
+Kappa statistic                          0.998 
+Mean absolute error                      0.0018
+Root mean squared error                  0.0278
+Relative absolute error                  0.3629 %
+Root relative squared error              5.565  %
+Total Number of Instances           125973     
+
+=== Detailed Accuracy By Class ===
+
+                 TP Rate  FP Rate  Precision  Recall   F-Measure  MCC      ROC Area  PRC Area  Class
+                 0.999    0.001    0.999      0.999    0.999      0.998    1.000     1.000     normal
+                 0.999    0.001    0.999      0.999    0.999      0.998    1.000     1.000     anomaly
+Weighted Avg.    0.999    0.001    0.999      0.999    0.999      0.998    1.000     1.000     
+
+=== Confusion Matrix ===
+
+     a     b   <-- classified as
+ 67296    47 |     a = normal
+    76 58554 |     b = anomaly
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+=== Re-evaluation on test set ===
+
+User supplied test set
+Relation:     KDDTest
+Instances:     unknown (yet). Reading incrementally
+Attributes:   42
+
+=== Summary ===
+
+Correctly Classified Instances       17808               78.9922 %
+Incorrectly Classified Instances      4736               21.0078 %
+Kappa statistic                          0.5928
+Mean absolute error                      0.2087
+Root mean squared error                  0.4369
+Total Number of Instances            22544     
+
+=== Detailed Accuracy By Class ===
+
+                 TP Rate  FP Rate  Precision  Recall   F-Measure  MCC      ROC Area  PRC Area  Class
+                 0.970    0.346    0.679      0.970    0.799      0.635    0.965     0.962     normal
+                 0.654    0.030    0.967      0.654    0.780      0.635    0.965     0.971     anomaly
+Weighted Avg.    0.790    0.166    0.843      0.790    0.788      0.635    0.965     0.967     
+
+=== Confusion Matrix ===
+
+    a    b   <-- classified as
+ 9421  290 |    a = normal
+ 4446 8387 |    b = anomaly
+
